refactor sample package structure
